Extract appChannelIsWaitingForFunding into helper
